requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
httpx==0.27.0
pandas==2.2.3
numpy==2.1.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0
finnhub-python==2.4.20
requests==2.32.3
tenacity==8.5.0
cachetools==5.5.0
nltk==3.9.1
pytest==8.3.0
black==24.10.0
ruff==0.7.0